reworked dynamicAttibute and limits properties of BaseItem
dynamicAttibute and limits used to be assignable by subclasses.
This causes the problem that these properties cannot be used in
qmldesigner and previous workaround is to create aliases, e.g.
    dynamicAttibute_channel
    limits_precChannel

Since qmldesigner requires these properties to be readonly, they
have to be initialized in the constructor.

But this fix also breaks those application created by qmldesigner.
There the appearance of "_" has to be replaced with ".".